+++ mlx_vlm/models/qwen2_5_vl/qwen2_5_vl.py
@@ -10,9 +10,9 @@ import mlx.nn as nn
 import numpy as np
 from huggingface_hub import snapshot_download
 
-from .language import LanguageModel, TextConfig
-from .vision import VisionConfig, VisionModel
-
+from .config import ModelConfig, TextConfig, VisionConfig
+from .language import LanguageModel
+from .vision import VisionModel
 
 @dataclass
 class ModelConfig:
